+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## neptune-client 0.15.2 [UNRELEASED]
 
 ## Features
+- Added support for workspace visibility in Management API ([#843](https://github.com/neptune-ai/neptune-client/pull/843))
 - Exposed container with a property of Handler ([#864](https://github.com/neptune-ai/neptune-client/pull/864))
 
 ## neptune-client 0.15.1